Ported ArithmeticParser example.
This example forced me to implement a couple of parsers, particularly:
- TrimmingParser
- CompositeParser
- DelegateParser
The last one is a little different than the original. In this case I'm just
using it as proxy that would delegate all its operations to another parser
that could be resolved any time in the future. The original implementation
made use of the Smalltalk #becomeForward: primitive, which C# doesn't have.